style: move base service to common
