+++ standalone/portal-api/src/common/services/base.service.ts
